+++ src/components/Header/styles.css
@@ -1,109 +1,75 @@
-.container {
-
-    width: 1143px;
-    height: 94px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
+.containerCustom {
+	width: 80%;
+	margin: auto;
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
 }
 
 .user-container {
-    width: 384px;
-    height: 84px;
-    display: flex;
-    align-items: center;
-    justify-content: left;
+	display: flex;
+	align-items: center;
+	justify-content: left;
 }
 
 .user-avatar {
-    width: 75px;
-    height: 69px;
-    border-radius: 50px;
-    margin-right: 17px;
+	width: 75px;
+	height: 69px;
+	border-radius: 50px;
+	margin-right: 17px;
 }
 
 .name-container {
-    font-family: 'Lato';
-    font-style: normal;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
+	display: flex;
+	align-items: center;
+	flex-direction: column;
 }
 
 .welcome-msg {
-    font-weight: 400;
-    font-size: 18px;
-    line-height: 24px;
-    letter-spacing: 0.1px;
-    color: #25282B;
-    margin-bottom: 8px;
+	font-weight: 400;
+	font-size: 18px;
+	color: #25282b;
+	margin-bottom: 8px;
 }
 
 .user-name {
-    font-weight: 700;
-    font-size: 40px;
-    line-height: 46px;
-    letter-spacing: 0.2px;
-    color: #52575C;
+	font-weight: 700;
+	font-size: 40px;
+	letter-spacing: 0.2px;
+	margin-bottom: 8px;
+	color: #52575c;
 }
 
 .user-title {
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    letter-spacing: 0.1px;
-    color: #52575C;
+	font-size: 16px;
+	color: #52575c;
 }
 
 .date-container {
-    width: 197px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+	display: flex;
+	align-items: center;
+	justify-content: center;
 }
 
 .calendar-icon {
-    width: 36px;
-    height: 40px;
-    color: #336CFB;
-    margin-right: 18px;
+	width: 36px;
+	height: 40px;
+	color: #336cfb;
+	margin-right: 18px;
 }
 
 .date-span {
-    height: 100%;
-    font-family: 'Lato';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 20px;
-    line-height: 26px;
-    display: flex;
-    align-items: center;
-    margin-top: 10px;
-    letter-spacing: 0.2px;
-
-    color: #25282B;
-}
-
-.button-container {
-    height: 40px;
-    width: 350px;
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
+	font-weight: 700;
+	font-size: 20px;
+	color: #25282b;
 }
 
 .logout-btn {
-    padding: 5px 25px;
-    background: #0156FF;
-    border-radius: 5px;
-    color: #FFFFFF;
-    font-family: 'Lato';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 16px;
-    line-height: 24px;
-    outline: none;
-    border: none;
+	padding: 5px 25px;
+	background: #336cfb;
+	border-radius: 5px;
+	font-weight: 700;
+	font-size: 16px;
+	outline: none;
+	border: none;
 }
